+++ ChestsAnywhere/Framework/Containers/ChestContainer.cs
@@ -38,7 +38,7 @@ namespace Pathoschild.Stardew.ChestsAnywhere.Framework.Containers
         public ContainerData Data { get; }
 
         /// <summary>Whether Automate options can be configured for this chest.</summary>
-        public bool CanConfigureAutomate { get; } = true;
+        public bool CanConfigureAutomate => this.Chest.SpecialChestType != Chest.SpecialChestTypes.JunimoChest;
 
 
         /*********